Layout Constraints Updated
The pages I have updated are sign up, sign in, caregiver, child, your plan, medications, allergies. Please only use an Iphone 15 Pro for the simulator when presenting to naren, I would need to deconstruct the entire views for medication, allergies, child, caregiver to have it fit every screen.
I also noticed some functionality is not working on these screens which is something to note on the trello board.