changes for merge oct 4
My changes are include storyboard
There is a segue that is mine, and a change to login.

Somehow the document segue name got overwritten, so if it doesn't work in QA testing it might have been overwritten again and I will need to merge this weekend again. Hopefully it works but if the change shows document view controlelr that was me!
Login view controller Storyboard: That was me since the textfield needs the secure password check